Pull out the use of Logger.new(), part of #8
This change initiates a Logging module which will hold the Ruby Logger.
Rather than having the instatiate the Logger for each class, we'll just
need to include this module.

As a result, any further changes to Logger that we may need to do, can
now be done from the Logging module.