Simplified the logger usage, moved core to Skewer::CLI, part of #8
The new logger usage relies on Skewer.logger rather than having to
explicitly include/require any files. However, in doing so it meant that
we can't have a Skewer class within the Skewer module as this messes
with the method search path.

To fix this, the Skewer::Skewer class has been moved to Skewer::CLI.