Clean up the interface for skewer, fixes #4
This change removes the nasty if-block that existed for the skewer
binary.

Added feature tests for the interface (high-level) to make sure that
we're providing sufficient help message when prompted, or when the user
provides bad input.

Also updated the way that Skewer::CLI::Parser handles failures. Rather
than raising exceptions, it instead calls the messages against abort().